requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
openai==0.28
python-dotenv